refactor(service/subgraph-ingester): prep for restful api implementation
