+++ services/ingest/BUILD.bazel
@@ -24,11 +24,11 @@ go_image(
 
 container_push(
     name = "push_image",
-    image = ":image",
     format = "OCI",
+    image = ":image",
     registry = "ghcr.io",
     repository = "z5labs/megamind/subgraph-ingester",
-    skip_unchanged_digest = True, # don't push if its the same
+    skip_unchanged_digest = True,  # don't push if its the same
     tag = "$(version)",
-    tags = ["manual"]
+    tags = ["manual"],
 )
