+++ services/ingest/cmd/BUILD.bazel
@@ -4,12 +4,14 @@ go_library(
     name = "cmd",
     srcs = [
         "cmd.go",
+        "grpc.go",
         "root.go",
+        "serve.go",
     ],
     importpath = "github.com/z5labs/megamind/services/ingest/cmd",
     visibility = ["//visibility:public"],
     deps = [
-        "//services/ingest/service",
+        "//services/ingest/grpc",
         "@com_github_spf13_cobra//:cobra",
         "@com_github_spf13_viper//:viper",
         "@org_uber_go_zap//:zap",
