+++ services/ingest/grpc/proto/BUILD.bazel
@@ -20,7 +20,16 @@ go_proto_library(
 
 go_library(
     name = "proto",
-    embed = [":service_go_proto"],
-    importpath = "github.com/z5labs/megamind/services/ingest/service/proto",
+    embed = [":proto_go_proto"],
+    importpath = "github.com/z5labs/megamind/services/ingest/grpc/proto",
     visibility = ["//visibility:public"],
+)
+
+go_proto_library(
+    name = "proto_go_proto",
+    compilers = ["@io_bazel_rules_go//proto:go_grpc"],
+    importpath = "github.com/z5labs/megamind/services/ingest/grpc/proto",
+    proto = ":service_proto",
+    visibility = ["//visibility:public"],
+    deps = ["//subgraph"],
 )
